Fix formatting of null values
Currently, null in an old value is formatted to U+002D HYPHEN-MINUS,
while null in a new value is formatted to U+2014 EM DASH. These are not
the same, so properties would show up as being changed if they were
null and unchanged. This fixes this by doing the formatting in one
place, thus standardizing on the U+2014 EM DASH option.